moved String instance
