+++ src_smartcontract_vm/CMakeLists.txt
@@ -11,6 +11,7 @@ add_subdirectory(instance_gc)
 add_subdirectory(instance_parts)
 add_subdirectory(instance_ref)
 add_subdirectory(instance_ref_static)
+add_subdirectory(instance_string)
 add_subdirectory(memory)
 add_subdirectory(reserved_classes)
 add_subdirectory(reserved_classes_string)
