+++ src_smartcontract_vm/instance/CMakeLists.txt
@@ -5,7 +5,6 @@ set(__src
 	IInstanceContainer.cpp
 	VmClassInstance.cpp
 	VmInstanceTypesConst.cpp
-	VmStringInstance.cpp
 )
 handle_sub(codablecashlib "${__src}" smartcontract_vm instance)
