+++ src_smartcontract_vm/instance_string/CMakeLists.txt
@@ -1,0 +1,7 @@
+
+
+set(__src
+	VmStringInstance.cpp
+)
+handle_sub(codablecashlib "${__src}" smartcontract_vm instance_string)
+
