+++ src_smartcontract_vm/instance_string/VmStringInstance.h
