+++ src_test/smartcontract_vm/reserved_classes/test_string_class.cpp
@@ -43,3 +43,8 @@ TEST(TestStringClassGroup, base01){
 	CHECK(result)
 }
 
+TEST(TestStringClassGroup, stringInst01){
+	UnicodeString str;
+
+}
+
